+++ CHANGELOG.md
@@ -8,6 +8,12 @@ next request, whatever version of [`@entur/uniformen`](./packages/uniformen) it 
 That is why entries have dates instead of version numbers. Changes to the package's
 API are listed in [its own changelog](./packages/uniformen/CHANGELOG.md).
 
+## 2026-09-24
+
+### Fixed
+
+- With `app=bedrift`, the logo links to `/bedrift` instead of `/`.
+
 ## 2026-09-23
 
 ### Added